Error fix 5 and add rwlock.rs

+++ src/FreeRTOS/src/FreeRTOSConfig.rs
@@ -3,7 +3,7 @@ macro_rules! configTICK_RATE_HZ { () => { ( 1000 as u32 ) }; }
 #[macro_export]
 macro_rules! configMAX_PRIORITIES { () => { ( 8 ) }; }
 #[macro_export]
-macro_rules! configMINIMAL_STACK_SIZE { () => { ( 200 as u16 ) }; }
+macro_rules! configMINIMAL_STACK_SIZE { () => { ( 200 as u64 ) }; }
 #[macro_export]
 macro_rules! configTOTAL_HEAP_SIZE { () => { ( 124 * 1024 ) }; }
